requirements.txt
streamlit>=1.20
python-dotenv>=1.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
chromadb>=0.5.3
sentence-transformers>=2.2.2
pydantic>=1.10.12
json-repair>=0.23.0

langchain-huggingface==0.0.3